projects/email-automation-bot/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.0
pymongo==4.6.0
redis==5.0.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
nltk==3.8.1
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
email-validator==2.1.0
python-multipart==0.0.6
jinja2==3.1.2
